fixed bug that edit user page viewed only on /calendar mapping
