+++ app/src/App.js
@@ -6,6 +6,28 @@ import Bar from "./components/bar/Bar";
 import {observer} from "mobx-react";
 import NewEventPage from "./components/newEvent/NewEventPage";
 import {userModel} from "./models/UserModel";
+import UserEdit from "./components/UserEdit";
+import Modal from "react-modal";
+
+const customStyles = {
+    overlay: {
+        display: 'flex',
+        alignItems: 'center',
+        top: 0,
+        bottom: 0,
+        right: 0,
+        left: 0,
+
+    },
+    content: {
+        overflow: 'hidden',
+        backgroundColor: 'transparent',
+        borderColor: 'transparent',
+        margin: '0 auto',
+        width: '50%',
+        padding: '0'
+    }
+};
 
 @observer
 class App extends Component {
@@ -16,18 +38,30 @@ class App extends Component {
 
     }
 
+    closeModal() {
+        userModel.userEditIsOpen = false;
+    }
 
     render() {
 
         return (
-            <Router>
-                <Switch>
-                    <Route path='/' exact component={Registration}/>
-                    <Route path='/login' exact component={Login}/>
-                    <Route path='/newEvent' exact component={NewEventPage}/>
-                    <Route path='/calendar' exact component={Bar}/>
-                </Switch>
-            </Router>
+            <div>
+                <Modal
+                    style={customStyles}
+                    isOpen={userModel.userEditIsOpen}
+                    onRequestClose={this.closeModal}
+                >
+                    <UserEdit/>
+                </Modal>
+                <Router>
+                    <Switch>
+                        <Route path='/' exact component={Registration}/>
+                        <Route path='/login' exact component={Login}/>
+                        <Route path='/newEvent' exact component={NewEventPage}/>
+                        <Route path='/calendar' exact component={Bar}/>
+                    </Switch>
+                </Router>
+            </div>
         );
       }
 }
